# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tahir AI Writer API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  author: "author",
} as const;

export interface UserPermissions {
  canGenerate: boolean;
  canPublish: boolean;
  canSchedule: boolean;
  canManageSites: boolean;
  /** @nullable */
  maxSites?: number | null;
  /** @nullable */
  maxArticlesPerDay?: number | null;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  permissions?: UserPermissions;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  author: "author",
} as const;

export interface CreateUserRequest {
  email: string;
  name: string;
  password: string;
  role: CreateUserRequestRole;
  permissions?: UserPermissions;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  admin: "admin",
  author: "author",
} as const;

export interface UpdateUserRequest {
  email?: string;
  name?: string;
  role?: UpdateUserRequestRole;
  permissions?: UserPermissions;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  newPassword: string;
}

export interface ResetPasswordRequest {
  newPassword: string;
}

export type SiteStatus = (typeof SiteStatus)[keyof typeof SiteStatus];

export const SiteStatus = {
  connected: "connected",
  disconnected: "disconnected",
  pending: "pending",
} as const;

export interface Site {
  id: number;
  userId: number;
  /** @nullable */
  folderId?: number | null;
  name: string;
  url: string;
  username: string;
  status: SiteStatus;
  isPinned: boolean;
  category?: string;
  /** @nullable */
  lastTested?: string | null;
  createdAt: string;
}

export interface CreateSiteRequest {
  name?: string;
  url: string;
  username: string;
  applicationPassword: string;
  /** @nullable */
  folderId?: number | null;
  category?: string;
}

export interface UpdateSiteRequest {
  name?: string;
  url?: string;
  username?: string;
  applicationPassword?: string;
  /** @nullable */
  folderId?: number | null;
  category?: string;
  isPinned?: boolean;
}

export interface BulkAddSitesRequest {
  sites: CreateSiteRequest[];
}

export interface BulkAddSitesResponse {
  added: number;
  failed: number;
  sites: Site[];
}

export interface AutoConnectRequest {
  url: string;
  /** @nullable */
  folderId?: number | null;
}

export interface SiteTestResult {
  siteId: number;
  success: boolean;
  message: string;
}

export interface TestAllSitesResponse {
  results: SiteTestResult[];
  totalTested: number;
  totalPassed: number;
  totalFailed: number;
}

export interface Folder {
  id: number;
  userId: number;
  name: string;
  siteCount: number;
  createdAt: string;
}

export interface CreateFolderRequest {
  name: string;
}

export interface UpdateFolderRequest {
  name?: string;
}

export type ArticleStatus = (typeof ArticleStatus)[keyof typeof ArticleStatus];

export const ArticleStatus = {
  queued: "queued",
  generating: "generating",
  published: "published",
  draft: "draft",
  failed: "failed",
  scheduled: "scheduled",
} as const;

export interface Article {
  id: number;
  userId: number;
  /** @nullable */
  siteId?: number | null;
  /** @nullable */
  siteName?: string | null;
  /** @nullable */
  siteUrl?: string | null;
  keyword: string;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  content?: string | null;
  language: string;
  /** @nullable */
  wordCount?: number | null;
  status: ArticleStatus;
  /** @nullable */
  publishedUrl?: string | null;
  /** @nullable */
  featuredImageUrl?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
}

export interface ArticleListResponse {
  articles: Article[];
  total: number;
  page: number;
  limit: number;
}

export interface ArticleUrl {
  id: number;
  title: string;
  url: string;
  publishedAt: string;
}

export interface SiteArticleUrls {
  siteName: string;
  siteUrl: string;
  articles: ArticleUrl[];
}

export type GenerateRequestImageSource =
  (typeof GenerateRequestImageSource)[keyof typeof GenerateRequestImageSource];

export const GenerateRequestImageSource = {
  pollinations: "pollinations",
  url: "url",
  none: "none",
} as const;

export interface GenerateRequest {
  keywords: string[];
  siteIds: number[];
  language?: string;
  wordCount?: number;
  model?: string;
  imageSource?: GenerateRequestImageSource;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  writingStyle?: string | null;
  includeInternalLinks?: boolean;
}

export interface GenerateResponse {
  queued: number;
  jobIds: number[];
}

export interface GenerateStatus {
  queueLength: number;
  processing: number;
  completed: number;
  failed: number;
}

export interface DashboardStats {
  totalSites: number;
  connectedSites: number;
  publishedToday: number;
  draftsToday: number;
  failedToday: number;
  totalArticles: number;
  totalPublished: number;
  totalFailed: number;
  totalDrafts: number;
  scheduledCount: number;
}

export interface SiteSummary {
  id: number;
  name: string;
  url: string;
  status: string;
  category?: string;
  articlesPublished: number;
}

export interface SitesOverview {
  totalActive: number;
  totalConnected: number;
  sites: SiteSummary[];
}

export interface CustomApi {
  id: string;
  name: string;
  baseUrl: string;
  apiKey: string;
  isDefault: boolean;
}

export interface AppSettings {
  longcatApiKey: string;
  longcatModel: string;
  pollinationsEnabled: boolean;
  defaultLanguage: string;
  defaultWordCount: number;
  concurrentGenerations: number;
  customApis: CustomApi[];
}

export interface UpdateSettingsRequest {
  longcatApiKey?: string;
  longcatModel?: string;
  pollinationsEnabled?: boolean;
  defaultLanguage?: string;
  defaultWordCount?: number;
  concurrentGenerations?: number;
  customApis?: CustomApi[];
}

export type ListSitesParams = {
  /**
   * @nullable
   */
  folderId?: number | null;
  search?: string;
};

export type ListArticlesParams = {
  status?: ListArticlesStatus;
  search?: string;
  /**
   * @nullable
   */
  siteId?: number | null;
  page?: number;
  limit?: number;
};

export type ListArticlesStatus =
  (typeof ListArticlesStatus)[keyof typeof ListArticlesStatus];

export const ListArticlesStatus = {
  all: "all",
  published: "published",
  draft: "draft",
  failed: "failed",
  scheduled: "scheduled",
} as const;