# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tahir AI Writer API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change own password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /folders:
    get:
      operationId: listFolders
      tags: [folders]
      summary: List folders for current user
      responses:
        "200":
          description: List of folders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Folder"
    post:
      operationId: createFolder
      tags: [folders]
      summary: Create a folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
      responses:
        "201":
          description: Folder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
  /folders/{id}:
    patch:
      operationId: updateFolder
      tags: [folders]
      summary: Update folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderRequest"
      responses:
        "200":
          description: Folder updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
    delete:
      operationId: deleteFolder
      tags: [folders]
      summary: Delete folder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles for current user
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [all, published, draft, failed, scheduled]
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: siteId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleListResponse"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /articles/{id}/retry:
    post:
      operationId: retryArticle
      tags: [articles]
      summary: Retry a failed article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Article queued for retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, author]
        permissions:
          $ref: "#/components/schemas/UserPermissions"
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - createdAt
    UserPermissions:
      type: object
      properties:
        canGenerate:
          type: boolean
        canPublish:
          type: boolean
        canSchedule:
          type: boolean
        canManageSites:
          type: boolean
        maxSites:
          type: ["integer", "null"]
        maxArticlesPerDay:
          type: ["integer", "null"]
      required:
        - canGenerate
        - canPublish
        - canSchedule
        - canManageSites
...
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required:
        - currentPassword
        - newPassword
...
    Folder:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        siteCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - siteCount
        - createdAt
    CreateFolderRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    UpdateFolderRequest:
      type: object
      properties:
        name:
          type: string
    Article:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        siteId:
          type: ["integer", "null"]
        siteName:
          type: ["string", "null"]
        siteUrl:
          type: ["string", "null"]
        keyword:
          type: string
        title:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        language:
          type: string
        wordCount:
          type: ["integer", "null"]
        status:
          type: string
          enum: [queued, generating, published, draft, failed, scheduled]
        publishedUrl:
          type: ["string", "null"]
        featuredImageUrl:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - keyword
        - language
        - status
        - createdAt
    ArticleListResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/Article"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - articles
        - total
        - page
        - limit
...