# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tahir AI Writer API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserPermissions } from "./userPermissions";
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  permissions?: UserPermissions;
  createdAt: string;
}